user BE structure change & email login update
